fast_api_tcc/pyproject.toml
[tool.poetry]
name = "fast-api-tcc"
version = "0.1.0"
description = ""
authors = ["Daniel Cruz <daniel.jack.dmc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.4"
alembic = "^1.13.2"
logging = "^0.4.9.6"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.1"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
httpx = "^0.27.0"

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.ruff]
line-length = 100
extend-exclude = ['Migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
run = 'fastapi dev app.py'
format = 'ruff check . --fix && ruff format .'
lint = 'ruff check . && ruff check . --diff'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fast_api_tcc -vv'
post_test = 'coverage html'

[build-system]
requires = ['poetry-core']
build-backend = 'poetry.core.masonry.api'
